--- src/components/upload-revenue/UploadHistoryTable/loading.tsx
@@ -1,15 +1,0 @@
-import classNames from "classnames/bind";
-
-import styles from "./loading.module.scss";
-
-const cx = classNames.bind(styles);
-
-const Loading = () => {
-  return (
-    <div className={cx("container")}>
-      <div className={cx("loader")} />
-    </div>
-  );
-};
-
-export default Loading;
